+++ recipes/pip.rb
@@ -1,4 +1,8 @@
 #
+# Cookbook Name:: thumbor
+# Recipe:: pip
+# Description:: Installs thumbor from GitHub repository
+#
 # Author:: Enrico Stahn <mail@enricostahn.com>
 #
 # Copyright 2012-2015, Zanui <engineering@zanui.com.au>
@@ -16,23 +20,7 @@
 # limitations under the License.
 #
 
-source 'https://rubygems.org'
-
-gem 'rake'
-
-group :test do
-  gem 'foodcritic'
-  gem 'rubocop'
-  gem 'serverspec'
-  gem 'infrataster'
-end
-
-group :integration do
-  gem 'berkshelf'
-  gem 'test-kitchen'
-  gem 'kitchen-vagrant'
-end
-
-group :releasing do
-  gem 'stove'
+python_pip 'thumbor' do
+  version node['thumbor']['version']
+  action :install
 end