+++ test/cookbooks/thumbor_test/metadata.rb
@@ -1,9 +1,0 @@
-# encoding: UTF-8
-
-name 'thumbor_test'
-maintainer 'Zanui'
-maintainer_email 'engineering@zanui.com.au'
-license 'Apache 2.0'
-description 'Installs/Configures thumbor_test'
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '0.1.0'